refactor(ui): fast-refresh compliance batch 6 — blocks.$ref.tsx + health.tsx
Part of #7850. 16 of 246 react-refresh/only-export-components warnings.

Same pattern as prior batches. blocks.$ref.tsx's BlockDetailPage reads
Route.useParams()/Route.useLoaderData() and renders <AppShell
crumbLabel={crumbLabel}> -- swapped to useParams({from})/useLoaderData({from}).
health.tsx's Incidents component reads Route.useSearch()/fullPath for its
status filter -- same useSearch({from})/useNavigate({from}) swap as
earlier batches. health.tsx additionally now exports `type HealthView`
and `type StateFilter` (previously local, needed by the moved page
component) -- type-only exports don't trip this rule.

detail-page-furniture-5481.test.ts source-scans blocks.$ref.tsx for the
literal `<AppShell crumbLabel={crumbLabel}>` string (that JSX call is
unreachable via rendering in its node-environment suite, same rationale
as validators-index-empty-action.test.ts) -- repointed its read target
to the new -blocks-ref-page.tsx sibling that JSX now lives in; all of
its assertions (including this one) still pass.

Zero behavior change; no external importers of any moved component
existed outside its own route file.

222 -> 206 warnings; 0 eslint errors. Live-verified /health (full
dashboard with live counts/incidents) and a block detail page (chain
walk, extrinsics/events tables, breadcrumb crumbLabel correctly showing
the loader-resolved block number) -- both render correctly, no console
errors (confirmed via a clean vite-error-overlay check, not just log
absence).
